requirements.txt
asgiref==3.6.0
backports.zoneinfo;python_version<"3.9"
Django==4.1.5
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
